Adds put_stream, compression and breaks lib.rs into modules
