+++ Cargo.toml
@@ -32,3 +32,4 @@ serde_json = "1"
 regex = "1.10"
 backoff = "0.4"
 hyper = { version = "0.14", default-features = false }
+async-compression = { version = "0.4.6", default-features = false, features = ["tokio", "gzip", "zstd"] }
